+++ plugins/dynamic-frontend-plugin/src/index.ts
@@ -1,1 +1,0 @@
-// dummy file for static build check