+++ plugins/jfrog-artifactory/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @janus-idp/backstage-plugin-jfrog-artifactory [1.3.6](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-jfrog-artifactory@1.3.5...@janus-idp/backstage-plugin-jfrog-artifactory@1.3.6) (2024-05-09)
+
+
+
+### Dependencies
+
+* **@janus-idp/shared-react:** upgraded to 2.6.1
+* **@janus-idp/cli:** upgraded to 1.8.6
+
 ## @janus-idp/backstage-plugin-jfrog-artifactory [1.3.5](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-jfrog-artifactory@1.3.4...@janus-idp/backstage-plugin-jfrog-artifactory@1.3.5) (2024-05-02)