+++ plugins/nexus-repository-manager/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @janus-idp/backstage-plugin-nexus-repository-manager [1.5.6](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-nexus-repository-manager@1.5.5...@janus-idp/backstage-plugin-nexus-repository-manager@1.5.6) (2024-05-09)
+
+
+
+### Dependencies
+
+* **@janus-idp/shared-react:** upgraded to 2.6.1
+* **@janus-idp/cli:** upgraded to 1.8.6
+
 ## @janus-idp/backstage-plugin-nexus-repository-manager [1.5.5](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-nexus-repository-manager@1.5.4...@janus-idp/backstage-plugin-nexus-repository-manager@1.5.5) (2024-05-02)